fix unittest timeout
